Fix mobile build (#2)
* Changed app name and bumped up min sdk version

* fixed angular prod build errors

These errors were found when the code was put onto an Android device. For some reason, the errors were not shown in the development environment but were prohibitive to running the app on the mobile device. Fixing this errors allowed the app to work as expected on the mobile device.